backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
opencv-python==4.8.1.78
fer==22.5.1
tensorflow==2.13.0
numpy==1.24.3
pydub==0.25.1
python-multipart==0.0.6
websockets==12.0
aiofiles==23.2.1
scipy==1.11.4
librosa==0.10.1
soundfile==0.12.1
pillow==10.1.0
matplotlib==3.8.1
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
pytest==7.4.3
pytest-asyncio==0.21.1